IntValue is now using int64 type. Implement time and rlIsKeyPressed operations. Actually begin working on the game.

+++ eval/evaluation_stack.go
@@ -24,7 +24,7 @@ func (stack *EvaluationStack) pushBool(b bool) {
 	stack.stack[stack.top] = NewBool(b)
 }
 
-func (stack *EvaluationStack) pushInt(i int) {
+func (stack *EvaluationStack) pushInt(i int64) {
 	stack.top += 1
 	stack.stack[stack.top] = NewInt(i)
 }
